requirements.txt
networkx
ujson
pandas
argparse
tqdm 
boto3 
requests 
regex 
sentencepiece 
sacremoses
nltk
lxml
bs4
torchtext
spacy==2.3.2
mosestokenizer
autocorrect
pytorch_pretrained_bert
transformers
pandarallel
gensim
easydict